Fixed errors reported by -Wextra

+++ sys/physmem.c
@@ -63,19 +63,19 @@ void pm_seg_init(pm_seg_t *seg, pm_addr_t start, pm_addr_t end,
 
   unsigned max_size = min(PM_NQUEUES, ffs(seg->npages)) - 1;
 
-  for (int i = 0; i < seg->npages; i++) {
+  for (unsigned i = 0; i < seg->npages; i++) {
     vm_page_t *page = &seg->pages[i];
     bzero(page, sizeof(vm_page_t));
     page->paddr = seg->start + PAGESIZE * i;
     page->vaddr = seg->start + PAGESIZE * i + vm_offset;
-    page->size = 1 << min(max_size, ctz(i));
+    page->size = 1 << min(max_size, (unsigned)ctz(i));
   }
 
   for (int i = 0; i < PM_NQUEUES; i++)
     TAILQ_INIT(PM_FREEQ(seg, i));
 
   int curr_page = 0;
-  int to_add = seg->npages;
+  unsigned to_add = seg->npages;
 
   for (int i = PM_NQUEUES - 1; i >= 0; i--) {
     unsigned size = 1 << i;
@@ -120,7 +120,7 @@ static vm_page_t *pm_find_buddy(pm_seg_t *seg, vm_page_t *pg) {
 
   intptr_t index = buddy - seg->pages;
 
-  if (index < 0 || index >= seg->npages)
+  if (index < 0 || index >= (intptr_t)seg->npages)
     return NULL;
 
   if (buddy->size != pg->size)
